move macos platform package
