+++ pkg/platform/bridge/library.h
