+++ pkg/platform/dummy.cpp
